feat: configure initial setup
