+++ votacao_frontend/src/app/(pages)/page.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <>
+      <h1 className="border border-amber-700 text-red-500">Hello</h1>
+    </>
+  );
+}
