+++ votacao_frontend/src/app/styles/globals.css
@@ -1,0 +1,17 @@
+@import 'tailwindcss';
+
+:root {
+  --background: #f1f1f1;
+  --foreground: #323c32;
+  --foreground: #323c32;
+  --color-principal: #33820d --color-secondary: ##64c832;
+}
+
+body {
+  background: var(--background);
+  color: var(--foreground);
+}
+
+h1 {
+  color: var(--color-principal);
+}
